requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1

# AI/ML Libraries
langchain==0.0.340
openai==1.3.7

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# API Integrations
slack-sdk==3.23.0
notion-client==2.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
python-multipart==0.0.6
requests==2.31.0
